+++ awslimitchecker/tests/services/test_directoryservice.py
@@ -1,5 +1,5 @@
 """
-awslimitchecker/tests/services/test_iam.py
+awslimitchecker/tests/services/test_directoryservice.py
 
 The latest version of this package is available at:
 <https://github.com/jantman/awslimitchecker>
@@ -33,13 +33,13 @@ either as a pull request on GitHub, or to me via email.
 ################################################################################
 
 AUTHORS:
+Di Zou <zou@pythian.com>
 Jason Antman <jason@jasonantman.com> <http://www.jasonantman.com>
 ################################################################################
 """
 
 import sys
-from awslimitchecker.tests.services import result_fixtures
-from awslimitchecker.services.iam import _IamService
+from awslimitchecker.services.directoryservice import _DirectoryserviceService
 
 # https://code.google.com/p/mock/issues/detail?id=249
 # py>=3.4 should use unittest.mock not the mock package on pypi
@@ -52,33 +52,29 @@ else:
     from unittest.mock import patch, call, Mock
 
 
-pbm = 'awslimitchecker.services.iam'  # module patch base
-pb = '%s._IamService' % pbm  # class patch pase
+pbm = 'awslimitchecker.services.directoryservice'  # module patch base
+pb = '%s._DirectoryserviceService' % pbm  # class patch pase
 
 
-class Test_IamService(object):
+class Test_DirectoryserviceService(object):
 
     def test_init(self):
         """test __init__()"""
-        cls = _IamService(21, 43)
-        assert cls.service_name == 'IAM'
-        assert cls.api_name == 'iam'
+        cls = _DirectoryserviceService(21, 43)
+        assert cls.service_name == 'Directory Service'
+        assert cls.api_name == 'ds'
         assert cls.conn is None
         assert cls.warning_threshold == 21
         assert cls.critical_threshold == 43
 
     def test_get_limits(self):
-        cls = _IamService(21, 43)
+        cls = _DirectoryserviceService(21, 43)
         cls.limits = {}
         res = cls.get_limits()
         assert sorted(res.keys()) == sorted([
-            'Groups',
-            'Users',
-            'Roles',
-            'Instance profiles',
-            'Server certificates',
-            'Policies',
-            'Policy Versions In Use',
+            'CloudOnlyDirectories',
+            'CloudOnlyMicrosoftAD',
+            'ConnectedDirectories',
         ])
         for name, limit in res.items():
             assert limit.service == cls
@@ -88,66 +84,69 @@ class Test_IamService(object):
     def test_get_limits_again(self):
         """test that existing limits dict is returned on subsequent calls"""
         mock_limits = Mock()
-        cls = _IamService(21, 43)
+        cls = _DirectoryserviceService(21, 43)
         cls.limits = mock_limits
         res = cls.get_limits()
         assert res == mock_limits
 
     def test_find_usage(self):
-        with patch('%s._update_limits_from_api' % pb) as mock_update:
-            cls = _IamService(21, 43)
+        mock_conn = Mock()
+        mock_conn.get_directory_limits.return_value = {
+            'DirectoryLimits': {
+                'CloudOnlyDirectoriesLimit': 3,
+                'CloudOnlyDirectoriesCurrentCount': 4,
+                'CloudOnlyDirectoriesLimitReached': True,
+                'CloudOnlyMicrosoftADLimit': 6,
+                'CloudOnlyMicrosoftADCurrentCount': 1,
+                'CloudOnlyMicrosoftADLimitReached': False,
+                'ConnectedDirectoriesLimit': 12,
+                'ConnectedDirectoriesCurrentCount': 0,
+                'ConnectedDirectoriesLimitReached': False
+            }
+        }
+        with patch('%s.connect' % pb) as mock_connect:
+            cls = _DirectoryserviceService(21, 43)
+            cls.conn = mock_conn
             assert cls._have_usage is False
             cls.find_usage()
-        assert mock_update.mock_calls == [call()]
+        assert mock_connect.mock_calls == [call()]
         assert cls._have_usage is True
-
-    def test_required_iam_permissions(self):
-        cls = _IamService(21, 43)
-        assert cls.required_iam_permissions() == [
-            'iam:GetAccountSummary'
-        ]
+        assert mock_conn.mock_calls == [call.get_directory_limits()]
+        assert cls.limits['CloudOnlyDirectories'].get_current_usage(
+        )[0].get_value() == 4
+        assert cls.limits['CloudOnlyMicrosoftAD'].get_current_usage(
+        )[0].get_value() == 1
+        assert cls.limits['ConnectedDirectories'].get_current_usage(
+        )[0].get_value() == 0
 
     def test_update_limits_from_api(self):
-        mock_summary = Mock(
-            summary_map=result_fixtures.IAM.test_update_limits_from_api
-        )
         mock_conn = Mock()
-        mock_conn.AccountSummary.return_value = mock_summary
-        with patch('%s.logger' % pbm) as mock_logger:
-            with patch('%s.connect_resource' % pb) as mock_connect:
-                cls = _IamService(21, 43)
-                cls.resource_conn = mock_conn
-                cls._update_limits_from_api()
+        mock_conn.get_directory_limits.return_value = {
+            'DirectoryLimits': {
+                'CloudOnlyDirectoriesLimit': 3,
+                'CloudOnlyDirectoriesCurrentCount': 4,
+                'CloudOnlyDirectoriesLimitReached': True,
+                'CloudOnlyMicrosoftADLimit': 6,
+                'CloudOnlyMicrosoftADCurrentCount': 1,
+                'CloudOnlyMicrosoftADLimitReached': False,
+                'ConnectedDirectoriesLimit': 12,
+                'ConnectedDirectoriesCurrentCount': 0,
+                'ConnectedDirectoriesLimitReached': False
+            }
+        }
+        with patch('%s.connect' % pb) as mock_connect:
+            cls = _DirectoryserviceService(21, 43)
+            cls.conn = mock_conn
+            assert cls._have_usage is False
+            cls._update_limits_from_api()
         assert mock_connect.mock_calls == [call()]
-        assert mock_conn.mock_calls == [call.AccountSummary()]
+        assert mock_conn.mock_calls == [call.get_directory_limits()]
+        assert cls.limits['CloudOnlyDirectories'].api_limit == 3
+        assert cls.limits['CloudOnlyMicrosoftAD'].api_limit == 6
+        assert cls.limits['ConnectedDirectories'].api_limit == 12
 
-        assert call.debug('Ignoring IAM AccountSummary attribute: %s',
-                          'GroupsPerUserQuota') in mock_logger.mock_calls
-
-        lim = cls.limits['Groups']
-        assert lim.api_limit == 100
-        assert lim.get_current_usage()[0].get_value() == 25
-
-        lim = cls.limits['Users']
-        assert lim.api_limit == 5000
-        assert lim.get_current_usage()[0].get_value() == 152
-
-        lim = cls.limits['Roles']
-        assert lim.api_limit == 501
-        assert lim.get_current_usage()[0].get_value() == 375
-
-        lim = cls.limits['Instance profiles']
-        assert lim.api_limit == 500
-        assert lim.get_current_usage()[0].get_value() == 394
-
-        lim = cls.limits['Server certificates']
-        assert lim.api_limit == 101
-        assert lim.get_current_usage()[0].get_value() == 55
-
-        lim = cls.limits['Policies']
-        assert lim.api_limit == 1000
-        assert lim.get_current_usage()[0].get_value() == 17
-
-        lim = cls.limits['Policy Versions In Use']
-        assert lim.api_limit == 10000
-        assert lim.get_current_usage()[0].get_value() == 53
+    def test_required_iam_permissions(self):
+        cls = _DirectoryserviceService(21, 43)
+        assert cls.required_iam_permissions() == [
+            'ds:GetDirectoryLimits'
+        ]